+++ NEWS.md
@@ -1,5 +1,8 @@
 # tidytable 0.6.2 (in development)
 
+#### Functionality improvements
+* `mutate.()`: Added `.keep` argument.
+
 # tidytable 0.6.1
 
 #### New functions